Fix audio glitch when starting/stopping HFP call
When audio is playing and HFP call is started
there can be observed multiple create/release
AudioPatch requests. When HFP call is started
and patch is created for non HFP audio source
it can be heard for few ms until patch
is released again.

In this approach connection will be invalidated
and connectAudioSource will be called only when
client primaryMix exists and its different
from strategy primaryMix or when client
lost its primary mix.

Bug: 351880230
Test: connect and disconnect HFP call
      when radio is playing.
Change-Id: I6fd89f7a9be6bc461df9db171df8f1afd75aa2f8