Change IAfEffectChain::sendMetadata_l to pure virtual
I encountered a "missing key function" error [1] for IAfEffectChain while experimenting with some compiler options, and the cause was that there is no definition for IAfEffectChain::sendMetadata_l. Because it does not appear that there is any intention of providing a default implementation in the base class, the appropriate fix for the error should be to make the function pure virtual instead.

[1]: https://lld.llvm.org/missingkeyfunction
Flag: EXEMPT refactor
Change-Id: I5ee49b88c07945e9b4fb58a0c81e5816a0902b9c