Check AID_RADIO correctly
AID_RADIO is just an appId. It has a different UID assigned
for each user on the device. When checking if the caller is
radio, extract the AID first and just check that.

This is a followup to some earlier targeted fixes to these
checks that partially/fully missed multiuser cases:

 - Ide5e9218811e110997fed19f43eb6a5fdc5812ec
 - I740b4f5acb88821197e3c4023a4032babd48ec5a
 - If5c721c9acfc78f1cbc35c3056adad95a4604f8b

This fixes a 10 second black screen caused by the dialer app
timing out while it waits, futilely, for media.metrics to become
visible to its process.

Repro as follows (must be on a secondary user):

 - adb shell am switch-user 10
 - adb shell pm clear com.android.phone
 - adb shell am start -n com.android.phone/.EmergencyDialer

Flag: EXEMPT trivial bugfix
Fix: 352390759
Test: Manual invocation of emergency dialer per above instructions
Change-Id: I2c8a6cc0459a6b46f482797e1456e30a70a43622