No fast mixer for DEEP_BUFFER threads
Do not initialize a fast mixer on output threads that is running as deep
buffer. Deep buffer is made for big buffers to save power and so on and
not for low latency which fast mixer is created for hence fastmixer is
not allowed on threads that is deep buffered.

Bug: 364329918
Test: manual
Change-Id: I256c168cde1f240a8c97473458bdb5aaaa09dc51