Mark android.hardware.drm@latest-service.clearkey as vendor
As part of the make to soong conversion, we are currently autogenerating
android_filesystem soong modules for each partition (system, system_ext,
vendor, ...), with the goal to eventually substitute the kati built
partitions. In order to generate the deps of these partitions, we are
currently using a heuristic to classify the modules in
`PRODUCT_PACKAGES` with the associated partition.

Mark the phony module `android.hardware.drm@latest-service.clearkey` as
`vendor: true` to help with this conversion heursitic.

Bug: 374371755
Test: verified that om.android.hardware.drm.clearkey.apex is present in the autogenerated soong-built
vendor partition

Change-Id: I4278e44b42b02877c46c66b3ddbe8e314369f270