Differentiate between mute reported to AS and port mute
The audio tracks now store the mute by port property which will be sent
to AS together with all the other mute properties. When changing the
muted by port we need to check whether the new muteState has changed.
This is why it is easier to have both properties stored within the
native tracks.

Test: atest AudioPlaybackConfigurationTest
Flag: EXEMPT bugfix
Bug: 374483735
Change-Id: Iba3153f13762bad3f055b0f69cf490e2ccff6d88