Block longer then timeout if first frame not drawn
The render thread repeats the last drawn frame when the virtual camera
owner fails to write a new buffer in time for the next capture request.
The time before the last buffer is repeated is define by the min fps and
the time of the last capture.

For the first frame, there is not last capture, so we need compute the
timeout differently. Also, to give time to the virtual camera owner
to setup the rendering for the first frame, we increase the allowed
waiting time for the first frame only.

Test: android.virtualdevice.cts.camera.VirtualCameraCaptureTest
Fixes: 364268484
Flag: EXEMPT Bug Fix
Change-Id: I302dc07b3cafc76f28d8aae763d6061472063afe