aaudio: fix SHARED MMAP streams
The HAL can return INVALID_OPERATION when the position is UNKNOWN.
That can cause SHARED MMAP to break. So coerce it to NOT_ENOUGH_DATA.
That will get converted to AAUDIO_ERROR_UNAVAILABLE,
which is ignored because it is OK if the DSP position
is temporarily unavailable.

Bug: 376467258
Test: Launch OboeTester
Test: Press TEST OUTPUT
Test: Disable "Exclusive"
Test: Press OPEN then START
Test: You should hear a sine wave.
Test: Also run the Data Path tests.
Flag: EXEMPT bugfix
Change-Id: Id46de86a6083dec50db93fb94c9c9cb34a7d8274