libaudiohal: Fix Standby sequence
If stream is in DRAINING state and standby is received then it is paused first. However, as DRAINING is a transient state for which HAL will be moved to IDLE on next cycle/trasient timeout. So, even after pause, stream can stay in IDLE state.

Allow reply.state to in IDLE after pause command (which is possible if original HAL state is DRAINING) so later flush,
standby sequence can be executed as part of intended fallthrough.

Bug: 377443079

Test: Oboe Tester -> open ->start -> stop, check logs for standby
Test: run cts-dev -m CtsNativeMediaAAudioTestCases

Change-Id: I882fd90c3bc49bfc40fecec169406b8ef88ecf2d