Audio CAP: Address ANAPIC comments, Part 3 (Split back AudioPolicyForceUse)
Revert "Migrate forced config and force use to android.media.audio.common.types"
This reverts commit 873738c761c34b56c018783455d908af1edf0605.

The reason is that the stable version of parcelables `AudioPolicyForceUse`
and `AudioPolicyForcedConfig` is being refactored for better
structure. However since in the framework these types are used up to
the Java layer, leave the framework versions as is.

Bug: 364310317
Test: m
Change-Id: Ib36958c042f4e52186fe9d5c31bae45473b2ebe4