Fix null pointer dereference in Camera2Client::stopPreviewL
stopPreviewL may in rare cases be called before initializeImpl. This is
highly unusual and should not happen, however when it does happen
crashing the server is not a good way to recover. Check if we've
initialized or not, and if not early out of topPreviewL given there's
no stream to stop.

Bug: 290086710
Test: Built locally
Flag: EXEMPT no significant behavior change
Change-Id: I19bf2e1737c507081be4d3222d532f46b539087c