Nuplayer: Handle get input/output format errors cleanly
In NuplayerDecoder, after MediaCodec::configure() returns success,
it is assumed that MediaCodec is in CONFIGURED state.
But, after MediaCodec::configure() has returned success, the codec
may be recycled by the mediaresourcemanager and its status will be
set to released.
So, errors from MediaCodec::getInputFormat() and getOutputFormat()
need to be handled cleanly instead of assert checks.

Bug: 379989075
Test: MediaPlayerTest
Change-Id: Iaaa03fb264c6b8ba7c784a2c00111c8a5e6153a7
Signed-off-by: guochuang <guochuang@xiaomi.corp-partner.google.com>