Fix getSupportedSchemes() in DrmHal.cpp
The original code returns the status of the first call to getSupportedSchemes() and ignores the status of the second call. This is incorrect because the second call may succeed even if the first call fails. The correct behavior is to return OK if either call succeeds, and to return the status of the first call if both calls fail.

Flag: EXEMPT bugfix
Bug: 380122010
Change-Id: I46054f93ec434d2c25943c3b02024f21397e3d0b