Fix waitForNextFrame method
EglSurfaceTexture::waitForNextFrame was using Surface::waitForNextFrame
but this method only work if the Surface is not parceled, which is never
the case.

This was causing virtual camera to always wait for the longest frame
duration (minFps) when set, even if the surface was written to.

The fix is to use Surface.onFrameListener and manually keep track of the
frame number and waiting.

Flag: EXEMPT bug-fix
Test: android.virtualdevice.cts.camera.VirtualCameraCaptureTest#all_requests_fulfilled_when_low_min_fps
Bug: 381886812
Change-Id: Ib1cc28cef904153165c141df6a479d3f24a5adbe