Do not set stream state as STOPPED when stream EOS is pending.
When doing offload streaming, clients can set offload end of stream to
indicate all data for current track is sent. The underlying AudioTrack
will be stopped to wait all data being played. When all written data is
played, the stream will automatically resume. That will only happen when
the stream is at STOPPING state and not STOPPED state. When receiving
the callback for offload stream end, the state can be set as STOPPED if
the client is not waiting for offload stream end.

Bug: 385218875
Test: atest AAudioTests
Flag: EXEMPT bugfix
Change-Id: I6bc2cc8c44a1ab8240e4b9c37074372077158d03