APV: Make workarounds for EncoderProfileLevelTest to pass
Current APV test files have extra 32 bits in the beginning of CSD.
Make workaround in the SW codec for the test to pass.
Once we have correct files we will revert this workaround.
Also, force bit_depth_minus8 to report 2. Once OpenAPV has the fix we
will rever this as well

Bug: 392977359
Test: EncoderLevelTest, EncoderProfileLevelTest
Change-Id: Idbf092701b8f10821623a9d4164f7460185d4ad8