Fix/skip libaudioclient tests on Automotive
Routing tests were failing because Automotive uses a different
policy for routing which affects the choice of devices. This is
the same problem that happened to some CTS tests. The decision
was to skip these tests on the automotive platform.

Another type of failure was occurring because tests assumed
existence of the SPEAKER device. However, on the automotive
platform the speaker may be implemented as a BUS device.
Change the test to use any output device.

Bug: 379596346
Test: atest audiorecord_tests audiorouting_tests \
            trackplayerbase_tests audiosystem_tests audioeffect_tests
Change-Id: Ied03cc9d689b7fde19f96d0ac723fbc5bf53f50d